Refactor
1. Reduce states and transition states
2. Merge all states and transition states into one file
3. Separate header and source files
