--- src/ego_states/ego_state_change_lane.cpp
@@ -1,15 +1,0 @@
-//
-// Created by jun on 7/24/17.
-//
-#include "../ego.h"
-#include "ego_state_change_lane.h"
-#include "../ego_transition_states/ego_transition_state.h"
-
-
-EgoStateChangeLane::EgoStateChangeLane() {}
-
-EgoStateChangeLane::~EgoStateChangeLane() {}
-
-void EgoStateChangeLane::onUpdate(Ego &ego) {
-  // TODO: add collision check here.
-}
