--- src/ego_states/ego_state_change_lane.h
@@ -1,27 +1,0 @@
-//
-// Created by jun on 7/24/17.
-//
-#ifndef PATH_PLANNING_EGO_STATE_CHANGE_LANE_H
-#define PATH_PLANNING_EGO_STATE_CHANGE_LANE_H
-
-#include "ego_state.h"
-
-class Ego;
-
-
-class EgoStateChangeLane : public EgoState {
-protected:
-
-  // constructor
-  EgoStateChangeLane();
-
-public:
-
-  // destructor
-  ~EgoStateChangeLane() override;
-
-  void onUpdate(Ego& ego) override;
-};
-
-
-#endif //PATH_PLANNING_EGO_STATE_CHANGE_LANE_H
