--- src/ego_transition_states/ego_transition_CS_to_CLL.cpp
@@ -1,9 +1,0 @@
-//
-// Created by jun on 7/29/17.
-//
-
-#include "ego_transition_CS_to_CLL.h"
-
-EgoTransitionCSToCLL::EgoTransitionCSToCLL() {}
-
-EgoTransitionCSToCLL::~EgoTransitionCSToCLL() {}
