--- src/ego_transition_states/ego_transition_CS_to_CLL.h
@@ -1,22 +1,0 @@
-//
-// Created by jun on 7/29/17.
-//
-#ifndef PATH_PLANNING_EGO_TRANSITION_CS_TO_CLL_H
-#define PATH_PLANNING_EGO_TRANSITION_CS_TO_CLL_H
-
-#include "ego_transition_FT_to_CLL.h"
-
-
-class EgoTransitionCSToCLL : public EgoTransitionFTToCLL {
-
-public:
-  // constructor
-  EgoTransitionCSToCLL();
-
-  // destructor
-  ~EgoTransitionCSToCLL() override;
-
-};
-
-
-#endif //PATH_PLANNING_EGO_TRANSITION_CS_TO_CLL_H
